goapp mac command refactored
