+++ drivers/mac/bundle.go
@@ -41,6 +41,8 @@ type Bundle struct {
 	Category Category
 
 	// Reports whether the app runs in sandbox mode.
+	// It is set to false by goapp if no signing identifier is provided when
+	// building.
 	Sandbox bool
 
 	// Reports wheter the app runs in background mode.
